Revert "merged from development"
This reverts commit b241fa022e7253fd51f1072c6b5fb179a8976673, reversing
changes made to f22d165e96571d6e3b088a2afff80f9febe2763a.

+++ core/extractor_ING_CREDIT.py
@@ -1,7 +1,6 @@
 import re
 from datetime import datetime
 import sys
-from decimal import Decimal
 
 from typing import Iterator
 
@@ -73,7 +72,7 @@ class ING_CREDIT(Extractor):
 
                     # <strong class=3D"expandable-value">
                     value = get_text_from_tag(row.find(class_='3D"expandable-value"'))
-                    value = Decimal(value.replace(',',''))
+                    value = float(value.replace(',',''))
 
                     result.append({"period_name":period_name, 
                                     "date":date,
